requirements.txt
asgiref==3.2.10
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.16
djangorestframework==3.12.4
flake8==4.0.1
gunicorn==20.0.4
idna==3.3
importlib-metadata==4.2.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.6.1
packaging==21.3
psycopg2-binary==2.8.6
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.8
python-dotenv==0.20.0
pytz==2020.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.3.1
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
zipp==3.8.0
